Changed ecdsa_curve to rsa_bits
When validating the terraform file, an error occurred: "Expected ecdsa_curve to be one of [P224 P256 P384 P521]". As the comment above states, AWS only supports RSA; the command was changed from expecting an ecdsa_curve to rsa_bits, which subsequently removes the error and allows validation to succeed.